+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- RPM build script for Fedora 41 ([fedora41.sh](./package_managers/fedora41.sh))
+- Distro-specific docker-based tests, to test in clean environments (Fedora 41)
+- arctis-manager: added --daemon-only option (mainly for test purposes, but can be used to disable the UI)
+
+### Changed
+
+- In case of device error during the identification process, if something goes wrong, reset the selected device and log the error
+
 ## [1.6.2]
 
 ### Added